requirements.txt
alembic==1.13.1
aniso8601==9.0.1
anyio==4.3.0
attrs==23.2.0
Babel==2.14.0
bcrypt==4.1.2
bleach==6.1.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
dnspython==2.5.0
email-validator==2.1.0.post1
Flask==3.0.2
Flask-BabelEx==0.9.4
Flask-Bcrypt==1.0.1
Flask-CKEditor==0.5.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
flask-mailman==1.0.0
Flask-Migrate==4.0.5
Flask-PageDown==0.4.0
Flask-Principal==0.4.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
Flask-WTF==1.2.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
logtail-python==0.2.10
Mako==1.3.2
Markdown==3.5.2
MarkupSafe==2.1.5
msgpack==1.0.8
openapi-client==1.1.7
passlib==1.7.4
pillow==10.2.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.33.0
requests==2.31.0
resend==0.7.2
rpds-py==0.18.0
six==1.16.0
sniffio==1.3.1
speaklater==1.3
SQLAlchemy==2.0.27
svix==1.20.0
types-Deprecated==1.2.9.20240106
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
urllib3==2.2.1
webencodings==0.5.1
Werkzeug==3.0.1
wrapt==1.16.0
WTForms==3.1.2
